+++ core/src/com/ray3k/skincomposer/dialog/DialogDrawablesFilter.java
@@ -193,6 +193,7 @@ public class DialogDrawablesFilter extends Dialog {
         
         
         stage.setKeyboardFocus(findActor("text"));
+        ((TextField) findActor("text")).selectAll();
         
         return this;
     }